requirements.txt
qutip==5.2.2
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
pyyaml>=6.0
scipy>=1.11.0
imageio>=2.31.0
pyscf>=2.2.0
scikit-learn==1.7.2
reportlab>=4.0.0
psutil>=5.9.0
pytest>=7.4.0
streamlit>=1.28.0
plotly>=5.15.0
joblib>=1.3.0
numpy-quaternion>=2022.1.1
Pillow>=10.0.0